fix: change migrations foreign keys unsignedBugInteger to unsignedInteger

+++ database/migrations/2023_07_04_000002_create_teams_table.php
@@ -17,7 +17,7 @@ class CreateTeamsTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('slug')->unique();
-            $table->unsignedBigInteger('game_id');
+            $table->unsignedInteger('game_id');
             $table->timestamps();
 
             $table->foreign('game_id')
