+++ database/migrations/2023_07_04_000004_create_tournaments_teams_table.php
@@ -17,8 +17,8 @@ class CreateTournamentsTeamsTable extends Migration
             config('leaguefy-manager.database.tables.tournaments_teams'),
             function (Blueprint $table) {
                 $table->increments('id');
-                $table->unsignedBigInteger('tournament_id');
-                $table->unsignedBigInteger('team_id');
+                $table->unsignedInteger('tournament_id');
+                $table->unsignedInteger('team_id');
                 $table->integer('stage_entry')
                     ->default(0);
                 $table->integer('rank')
